Consolidated iterators under "util.collection" package
