+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/table/format/ParquetSplitRecordIterator.java
@@ -18,7 +18,7 @@
 
 package org.apache.hudi.table.format;
 
-import org.apache.hudi.common.util.ClosableIterator;
+import org.apache.hudi.common.util.collection.ClosableIterator;
 import org.apache.hudi.exception.HoodieIOException;
 import org.apache.hudi.table.format.cow.vector.reader.ParquetColumnarRowSplitReader;
